Refactored to test scheduler and TestSubProcess
The test scheduler and the TestSubProcess have been refactored and the
schedulers readability has greatly improved.

To increase readability, the TestSubProcess class now has defined 'status' that
it can bee (see the requirements document). It also has functions to check on
its status so the scheduler only has to query these functions, which reduces
the amount of logic in the scheduler.

Stubs and sections have also been created for future parts of the scheduler
that are not yet implemented.