Tests can communicate with the TestRunner
Tests can now communicate their results and status with the TestRunner.

Unfortunately, Python does not have many capabilities for communication between
process.

One of the main challenges for this project is to allow the test writer to not
have to worry about interprocess communication. The user shouldn't have to call
any methods for process communication, it should just be handled for them.

Threading is one opportunity for sharing dating between two processes, but
because of the Python's Global Interpreter Lock (GIL), any implementation using
Python threading wouldn't be multi-core.

The multiprocessing module also provides Pipes and Queues for process
communication. This method might actually be another solution. The
TestSubProcess could communicate between the test object and the test runner
both before the call to the test's run() function. This additional method for
communication should be examined.

The multiprocessing.shared_memory module could also provide a method for
communication between processes; however, it only available after 3.8. Most
installations of Python (for instance on Cheyenne) are <= 3.7.x.

The last interprocess communication that's provided by Python's multiprocessing
module are the Managers. A manager object controls a server process which
managers a shared object, while other processes can access the shared object by
using a proxy. This is the method that this commit implements. This is the
method (although extended in this implementation) used by the first version of
SMARTS.

Using a manager object and proxy object enables the TestRunner to keep track of
its status and results while it is running. This advantage will need to be
explored, it may not be necessary.

This implementation essentially makes each test class a 'shared object'. The
shared object is owned and controlled by the manager, which is in the test
runner and its proxy is given to the test itself. Thus, a change to the test
updates the proxy, which updates the shared object which the test runner can
see.

There maybe a loss of usability using this method. If a test contains a syntax
error, its error message and stack trace may contain references to the
multiprocessing class and not to the test itself.