Added base reporter
Added `reporter.py` within in smarts/reporters/. `reporter.py` defines
ResultClass, ResultProxy and Result. Using the multirproceccing managers,
reporter.py creates a shared object, Result, that will be passed to each test
to communicate the result.

The ResultProxy, which is derived from the multiprocessing NamespaceProxy,
enables the ability to access (and create) attributes of the ResultClass. That
way, tests can set the result attributes (and create their own) and allow the
TestRunner or other Reporters access.