Updated environment list_modsets, get_modset, and load_modset
list_modsets now returns the NAMES of the modsets and not the modsets
themselves. These names can be used in get_modset, which returns the modset,
but does not load it, and load_modset. Load_modset now requires the name of the
modset and will return the modset once it has been loaded successfully.