smarts.py removes duplicated requested tests
Previously, if a test was requested to be ran twice, they both would be
scheduled to run, and the second would fail as it could not create a test
directory for itself (because a directory already existed from the first test).

This commit fixes the above by changing smarts.py command line API to remove
any duplicated requested tests before passing it in to the TestRunner.

This method doesn't provide the user with any notification that a test was a
removed, or which one, thus a future commit might want to look into doing some
kind of reporting.