Implemented env._load_library
This commit implements the ability to load libraries of a modset. It implements
both loading the library as a lmod module and as a environment variable, value
pair.

This commit was only tested with loading libraries with the lmod command, so
the environment name, value pair option may not work.