Update error message for invalid and unfound
This commit updates the error message when a test cannot be loaded. In the
past, SMARTS would only print that the test was invalid, but not provide any
infromation on why it was invalid.

Now, the message will print the reason why a test was invalid, if the test can
be found, but not be loaded correctly. If the test cannot be found, then an
error message will be printed saying that it cannot be found in the test
directory.