env.parse_file checks for required section
This commit adds basic error checking on the environment.yaml file that was
passed in. In this commit, it checks if the required sections Description or
Modsets are missing. If it is missing, env.parse_files prints and error message
and returns -1.

Because the Environment class and this function may not necessarily be invoked
by a terminal, a future commit or version should instead return the error
messages rather then printing out the error message. That way each individual
API can handle the error and the error message in a method that they choose.