Add env.yaml error checking for description section
This commit adds in error checking for the missing or mismatched attributes of
the description section. Currently, two possible configurations will produce
warnings instead of errors: if the LMOD_CMD attribute is specified and the
Modules attribute is set to False, or is missing.

Currently, these warnings will cause SMARTS to exit, but a future commit will
enable them to continue running with LMOD support turned off.

It also adds in variety of erroneous env.yaml files for testing.