Handle warning for incorrectly specified lmod cmds
This commit handles the warnings for incorrectly specified lmod cmds. A warning
is produced if the LMOD_CMD was given, but no 'Modules' attribute was given. If
'Modules' is set to False, not warning is produced and LMOD_CMD is turned off.

If the LMOD_CMD is given, and no 'Modules' attribute was specified, then a
warning is produced alerting the user that LMOD support is turned off.