Add error checking for modset section of env.yaml
This commit adds in error checking for the env.yaml file. Similar to the
commits that came before it, it checks to see if basic parts of Modsets are
present. It also checks for inconsistencies across the file. For instance, if
the env.yaml Description specifies that "Modules: False" and a modset uses
"module" to specify a library, an error will be produced.