Capture and report lmod load errors
This commit captures and reports errors that may occur when `lmod python load`
encounters and error, for instance if lmod was given an unknown module name.
In this commit, if an error is encountered, the test process will call
sys.exit(-1), which will cause the test to fail (Assuming result.result is set
to None).

A future commit will, instead, return an error code from `env._lmod_load`, which
can be passed back to the user test and handled explicitly.