Percolate lmod errors up to tests
If a `lmod python load` command encounters and error, percolate the failure
back up to the test. At this point, only the fact that the command failed,
i.e., `False`, is communicated.

A future PR could percolate an error message, or a status code to communicate
the reason or location for failure.