Automatically load a tests dependents
This commit adds logic to TestManager.run_tests to automatically load
dependencies of a specified test (if they are not already specified to run).
This change will allow a user to specifiy a single test and have all of its
dependents be ran.

If a test dependent itself has any depedents, those too will be loaded and ran.