Fix incorrect output when test returned no result message
This commit fixes a bug which prevented the testManager to print out tests when
a specified test returned a NoneType for its result.msg. A test can return a
result.msg when it incounters a Python runtime error that it does not
explicitly handle/trap/etc or when a test fails to set the result of a test at
all (as the default of the result class attributes are all None).

Instead of concatinating the test result messages together, this commit uses
str.format() as it is more readable (as well, it ensures that if a None type is
attempted to be written (although it is caught before the write, that it will
be written as None, as the NoneType has a string representation None.__str__().

It may be prudent to consider changing the defaults of the result class, but
the testManager should still print out an ERROR for any test that fails to
correctly set the result class.