Implement ability to pass CL options via config files
This commit enables the ability for users to pass command line options to
SMARTs via configuration files.

Configuration files can be in either of two places:

1. User directory `~/.smartscf`
2. Current working directory `./.smartscf`

They allow users to pass the same options that can be passed via the command
line, the environment file, the test directory, the source directory and the
verbosity level.

On top of these files, the command line options can still be passed.

Smarts will try and read all of the above configuration files if they are
present. Precedents for options will be taken in the following order
(1 = highest priority):

1. Command line options
2. Current working directory `./.smartscf`
3. User config `~/.smartscf`